Refining Task-Action design for fleet adapters
