+++ rmf_fleet_adapter/CMakeLists.txt
@@ -65,6 +65,7 @@ target_link_libraries(read_only
 add_executable(full_control
   src/full_control/main.cpp
   src/full_control/FleetAdapterNode.cpp
+  src/full_control/MoveAction.cpp
 )
 
 target_link_libraries(full_control
